Add compatibility for closerange
Additionally, add a cpython version of closerange; while it's faster,
the extension exists so where possible we can avoid generating a massive
number of close syscalls (crapping up strace logs).  We could do the
equivalent in native python, but unfortunately it's fairly slow- thus
we do it this way.

Finally, monkey patch os.closerange with our extension if it's available.
This is done to keep subprocess.Popen calls quiet in addition.