Return a command class even when sphinx is failing.
Specifically, we return a command class that fails if invoked.  This
is preferable to the current form where None is returned, requiring
any consuming code to basically hide the commands if sphinx isn't
available.

Thus for previous code, `setup.py build_docs`, if sphinx was missing
was 'command not found'; after this change it's 'sphinx isn't available'.
Generally friendlier, and more useful (plus simpler for consuming code
to work with).