remove hash imports from outdated libs (mhash, fchksum, etc)
Both mhash and fchksum aren't py3 compatible and so aren't worth using
as a fallback going forward. In addition, the md5 and sha modules both
were dropped in py3 in favor of hashlib.

Now we use hashlib by default and fallback to pycrypto if necessary.
Note that the related hashing comments and benchmarks will be updated
later.