demandload: drop the initial required scope argument
This makes the API slightly nicer since we don't need to pass the
caller's global namespace as the first argument for all uses. If we
happen to find a case where we need to pass a different scope we might
re-add the scope argument as an optional arg or add a different method.

Note that we determine the caller's global namespace by going back one
frame in the stack and looking at f_globals.

We're taking liberties here since pkgcore hasn't had a release in a long
time and I doubt we have any other external users making use of
demandload specifically. Generally we'd make the scope argument optional
and deprecated for a while before dropping it after a few releases so
people running pkgcore wouldn't hit breakage.