test: skip mount permission failure test if running as root
The mount syscall will be allowed as root so no exception will be
raised. This then causes issues during the teardown phase that tries to
remove the directories and fails when the bind mount still exists.

We could re-enable it once a similar umount() is added that would
unmount the tempdir bind mount before removing the directories during
the teardown phase.