vastly simplify snakeoil.xml
Drop most of the fallbacks from trying to find a compatible elementtree
module. Just check for the C implementation and then fall back to the
native python implementation as necessary. In python3, using
cElementTree is deprecated as well so this module can be removed once we
drop python2 support.

Also, instead of our own internal method the escape method from
xml.sax.saxutils can be used.