osutils.ensure_dirs(): do not alter permissions on existing dirs
Remove the code attempting to add u+wx permission on existing dirs for
mkdir() calls. Fiddling with permissions in system directories (even
temporarily) is a bad idea, and in most cases it is unnecessary (root
bypasses permission checks, and non-root will fail anyway due to chown()
calls). It is better to fail when mkdir() would fail instead.