move packaging, testing, and distribution stuff into its own module
This forces a bit of code duplication between projects but the upside is
we don't have to install these non-essential bits and also setuptools
should work as expected while parsing deps from setup scripts due to
dropping the snakeoil dep from downstream projects, e.g. pkgcore and
pkgcheck.

Currently we just subtree it in as that's easiest for pushing/pulling
changes.