snakeoil.process.is_running(): Do not attempt to read /proc on non-Linux
Do not attempt to read /proc/*/status in is_running() when running on
non-Linux system. Non-Linux systems may either do not have /proc
filesystem (in which case the function wrongly returns ProcessNotFound)
or it is incompatible. In case of Solaris, the status file is binary
which results in UnicodeDecodeError.

Instead, assume the process is running if the PID is valid. However,
the long term goal should be to remove this function as unreliable
and use complete process control.