test/test_cli: fix argparse doc tests no matter what module import order
Previously, snakeoil.cli was getting imported by another test that scans
the entire module namespace and imports all the modules to perform some
action. If this happened before the initial cli test relying on an
unpatched argparse module, the test would fail.

Now we first reload the argparse module to drop the patches to perform
the tests requiring a clean module, then reload the snakeoil.cli module
to patch argparse and perform the remaining tests.

Fixes #28.