bash: add support for respecting line continuations in iter_read_bash()
In other words, lines that end in a \<newline> are continued on the next line,
properly respecting comments of course.

It's not enabled by default, because PMS specifies supporting single
line format only in profiles, but there isn't such a restriction for
system/user config files.

Runtime-wise, when allow_line_cont is disabled the function still runs
at about the same performance as previously. When allow_line_cont is
enabled it's about 2x slower due to the regex usage. Simpler line
continuation support could be implemented to speed things up, but it's
fine for now as it's currently only enabled when parsing
/etc/portage/package.use and similar settings.