cli.arghparse: ArgumentParser: make sure the main func is overridden properly for subcmds
Previously in certain cases the main func for the core parser would be
used instead even when running a subcommand due to splitting up argument
parsing for some reason. Now, assuming that the parent parser's
parse_args() method is always called (and it should be) the main func
should always be set properly in the namespace.