contexts: SplitExec: initial support for passing back local scope var changes
From the child process back to the parent process (inside the context
manager to outside).

Currently a simple, best effort transfer occurs where every changed
object in locals() that can be pickled with the pickle module is passed
back. This are made available via the locals attr on the context manager
obj. They aren't injected back in dynamically to the actual variable
names since locals() are readonly except under global scope and working
around that is incredibly hacky.