Fix `setup.py develop` support.
The scripts parameter is supposed to be a usable path, but
was just the basename of the final scripts.  This wasn't
visible due to `build_scripts.finalize_options()` rewriting it
to a proper path.

The `develop` command, however, has no usage of `build_scripts`
thus doesn't call it.  It instead pukes that the files in
the scripts param couldn't be found.

The fix is to use proper/absolute paths, and have build_scripts
handle that accordingly.

Additionally, while I'm in here, I'm fixing some issues in
`build_scripts` (option finalization creating directories, hardcoded
paths to lookup, etc).