test/mixins: (py3.11) fix invalid call to __subclasses__
In python3.11, a new metaclass and some classes were added into ctypes.
The metaclass `ctypes._endian._swapped_struct_meta`, while hidden, is
inheriting from `type(Structure)`, which result in a lot of magic inside
but at the end we get a state that it's `__subclasses__` is the same as
`type.__subclasses__` function, meaning that it expects an arg. This
metaclass didn't exists before 3.11.

In this commit you can see that I changed the `if` line to check the
signature of the function, and only if it doesn't expect an arg,
continue. Yes, the new `if` is slower, but note that this is only for
testing code, so we can live with it.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>