History: add rev suppression for commits that bulk moved files.
The point here is to be able to pull the full history for a change
rather than losing the thread because folks moved files around.  Some
of this code is old enough to vote, so knowing why something was done
back in py2.2 era matters.

This is also useful for un-obscuring some of the original authors for
double checking no major holder is being dropped.

Signed-off-by: Brian Harring <ferringb@gmail.com>