tests: Loosen the argparse exception check
Loosen the argparse exception check to fix tests with Python 3.14.
Starting with Python 3.14, the exception is supposed to use single
quotes rather than double quotes.  However, due to
https://github.com/python/cpython/pull/134971 it currently does not
include the name at all -- an easy way to get compatibility for both
quoting variants and to work around the bug is to not check the name
at all.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Part-of: https://github.com/pkgcore/snakeoil/pull/105
Closes: https://github.com/pkgcore/snakeoil/pull/105
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>