feat: saner Immutable metaclasses, and combine_classes helper.
combine_classes is just a tool for combining classes on the fly.  This
```
class foo: pass
class foo2: pass
class combined(foo, foo2): pass
```
is this:
```
combined = combine_classes(foo, foo2)
```

Use the former syntax for normal classes unless you have reasons to be terse;
for metaclasses however, for simple combinations just use combine_classes
since python metaclass usage makes this ergonomically a PITA.

Imutables:

The previous immutable implementation was 'strict' in requiring all mutation
to use object.__setattr__, which lead mutation blocks- __init__ primarily-
being painful.  The old implementation is wrapped into a proper type instance
to allow composition, and the old code (and this metaclass) is written like this:
```
class foo(metaclass=ImmutableStrict):
  def __init__(self):
    object.__setattr__(self, 'x', 1)
    self._subinit()
    self._subinit2()
  def _subinit(self): object.__setattr__(self, 'y', 1)
  def _subinit2(self): object.__setattr__(self, 'z', 1)

f = foo()
assert None is f._subinit()
assert None is f._subinit2()
```

The new Immutable class addresses all of that, which also makes IDE and static
analysis actually work sanely.
```
class foo(metaclass=Immutable):
  def __init__(self):
    self.x = 1
    self._subinit()
    self._subinit2()

  def _subinit(self): self.y = 1

  @Immutable.allow_mutation
  def _subinit2(self): self.z = 1

f = foo()
pytest.raises(AttributeError, f._subinit)
assert None is f._subinit2()
```

This is done via contextvars tricks, thus is both thread and async
safe.  The logic is such that when a mutation block is entered for that
instance, the instance is allowed to be mutated until returning from the
block (technically the context of the underlying contextmanager).

If during a mutable context, an invocation occurs that would enter another
mutable context for another instance, then it *does* shift context.  It's
both re-entrant and locked down.

Due to python generators not having their own context within that frame,
the generator will run under whatever context invokes or consumes from it.  An
implementation can be added to fix this, but that's ugly as hell; it'll require
wrapping the generator in a generator that switches contexts on the fly around yields.

If needed, I can write this, but it's extremely corner case, and if python ever
fixes the generator context problem, there are tests to catch this so the implementation
can be updated.

The pre-existing immutable APIs have all been marked as deprecated; users should migrate
to the new ImmutableStrict (no change in results), or Immutable if they want saner code.

Signed-off-by: Brian Harring <ferringb@gmail.com>
Closes: https://github.com/pkgcore/snakeoil/pull/108
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>