feat: add tooling for working with slotted klasses and instances.
This fixes something I noticed in klass.ProxiedAttrs and
klass.SlotsPicklingMixin that's edge case, but is a thing.  To
my knowledge this bug has never surfaced- if it ever came up,
it would've only been during debugging.  No code- to my knowledge-
ever hit this edge case.

Meanwhile, the particulars: unfortunately dir() isn't usable
for this level of inspection- it pickups property, methods,
class underlying objects, etc.

For slotting, one can't just walk __slots__.  Each layer
of inheritance can set this creating actual slots, and
this does not mean the cumulative (last) __slots__ is
total.  It's just whatever was last set.

Thus this; walk the MRO and extract out all possible
attribute names, and then return it.

The usage of this pretty much comes down to serialization,
and in adding this it fixes SlotsPicklingMixin to
address the MRO crap I mentioned above.

Signed-off-by: Brian Harring <ferringb@gmail.com>