chore: add deprecation suppression, pull in deprecated dependency
This change adds a runtime dep of the 'deprecated' package- this
https://pypi.org/project/Deprecated/ .

However, being that snakeoil can be used by critical code, the
internal snakeoil usage of it explicitly can fallback to py3.13
deprecated implementation, and if it's <py3.13 it just suppresses
the deprecations in full.

By "can fallback", this includes ImportError- broken import, and missing.

It's not optimal, but this is sane enough frankly.

Signed-off-by: Brian Harring <ferringb@gmail.com>