feat!: turn down slot_shadowing test defaults
This was from early days of limited memory and slow python, and
when snakeoil/pkgcore/pkgcheck releases were tightly coupled.

These days, there's no reason to explode if a slot_shadowing is
found- instead if should only go 'boom' if it's explicitly told
to do so.

For snakeoil, this is enforced since our dependency chain is
minimal.  For anything downstream, this isn't unless they enable
it.

Note that strictly speaking, this breaks semver behavior for
tests- what would've failed before now won't.  Downstream
consumers are all clean so this is a non issue however.

Signed-off-by: Brian Harring <ferringb@gmail.com>