fix: time format handling in git log parsing.
Date formats differ across countries, this code was written
to american norm due to the underlying git invocation being
sensitive to the users configuration for this.

The datetime parsed epoch is available for code to use for
localization, but for compatibility sake, this code still
returns the date formatted in american norms.

Signed-off-by: Brian Harring <ferringb@gmail.com>