chore: fix ABC of Formatter and add annotations.
Formatter base class requires .stream, so just force it.

I need to go in and just collapse that ontology, the annotations
are layer violations are painful AF due to the history of how
that evolved and tooling back then being insufficient to do this
sort of refactor safely.  This will be addressed via #113.

Signed-off-by: Brian Harring <ferringb@gmail.com>