fix: ManCoverter assumed directory structure that wasn't guaranteed.
I frankly do not know why this hasn't shown elsewhere, but I can repeatedly
introduce this on my localhost.

Being I flushed this out via removing some scripts from pkgcore, I'm assuming
there is some mapping in the guts of this and via removing a script, I changed
the execution order.  Whatever it was processing first prevented it being visible.

Either way, read the code: this is a clear bug.

Signed-off-by: Brian Harring <ferringb@gmail.com>