feat: implement code_quality.Modules for __all__ validation
I'm unclear how it got broke, but the ExportedModules class
does *nothing*, so just turn it into a deprecated noop.  It
literally doesn't run in any usage of it.

This is a rewrite of it that's saner and parameterized and
defaults to failing on __all__'s referencing non existant
exports.

Signed-off-by: Brian Harring <ferringb@gmail.com>