feat(test): make test.protect_process usable
This was adhoc working for formatters due to a quirk of snakeoil
module layout; instead, since the code uses pytest, just exec
the fork as a pytest call w/ the test that is wrapped.

This is now generally usable, although if you're using the decorator
within a test (not 'for the test'- within)- some care is warranted
since effectively that entire scope gets ran both on the fork and
parent.  See 'test_failure' for an example of how to handle this.

Signed-off-by: Brian Harring <ferringb@gmail.com>