feat(python_namespaces): add import_module_from_path for arbitary imports, and protect_imports for bad ideas.
protect_imports:  Use this to protect and unwind changes to sys.modules
and sys.path.  It is explicitly not thread safe because it's impossible
for it to be thread safe.  Generally this should only be used in tests,
but there are special cases where temporarily importing a secondary tree
has valu.  For example to grab a submodule within a tree- just that submodule-
and then discard everything but it, ensuring that there is no conflicting
intersection from that tree.  Again, useful for tests, but for plugin
architectures this has a fair amount of value.

import_module_from_path: use this for code that must import python source
that isn't part of a normal python tree, or has strong reasons
to ensure it doesn't get integrated into sys.modules.  It's edge
case.

Signed-off-by: Brian Harring <ferringb@gmail.com>