chore: set 0.11.0 as version, add pytest-subtests as a dep.
pytest-subtests can be removed once >=pytests-9.0 is allowed.

0.11.0 is marked since this is not 0.10 lineage; a ton of
code is getting deprecated.  Pkgcheck alone now has ~.5m
warnings thrown for example, which will be addressed as
pkgcore next gets updated to remove the deprecations, and
then pkgcheck.

There should be no semver breakage in 0.11.0 with the caveat
that code that was broken before- not implementing a class
fully for example- now *will* fail during execution.  It
should have always, but we've integrated better tools into
the code.

All downstream tests pass against 0.11.0, so again, semver
should not broken.  But if it is, your code was doing something
it shouldn't have even for 0.10.x.

Signed-off-by: Brian Harring <ferringb@gmail.com>