feat: deprecated can now attach metadata to use for tracking removals.
The point of deprecations should not just be "tell downstream devs"-
it is explicitly metadata within the codebase that can be leveraged
to track if there is deprecated code still in the codebase that is
past it's expiry date.

This does that, and is designed such that each codebase creates it's
own `deprecated` Registry which keeps their deprecations internal to
them.

In the follow on commits I'll add logic to leverage this metadata
and update the codebase to enrich the existing deprecations.

Finally: store this in _internals.  This isn't intended for public
consumption, thus put it there.  Other packages- pytest for example-
use _pytest in the global scope.  We can do so, I'm just starting
the work via keeping it in our package namespace for now.

Signed-off-by: Brian Harring <ferringb@gmail.com>