feat: add test.AbstractTest since pytest silently ignores ABC test classes.
Pytest intentionally ignores all classes that are still abstract during
collection.  This is entirely valid when one is dealing in intermediate
base classes, but there is no mechanism to say "I actually did want
it to be concrete".  If you make any mistake in the chain of subclassing
where you didn't override an abstract for a templated test you're adding,
pytest drops it.

This is an abc.ABC base for tests that requires explicitly marking
subclasses as abstract if that's the intention.  If they're not marked
but are still abstract, it throws a TypeError so people writing tests
see it, and resolve it.

Without this you can have tests silently dropping out of your suite.

It's a pain in the ass, even if I understand why they did what they did.

Signed-off-by: Brian Harring <ferringb@gmail.com>