feat: add collector for expired deprecations
Also add the ability to drop a note into the code for
things that must be removed.  A general code directive..

For example, deprecation.Registry itself has code that
must be updated in python >=3.13.  Thus add a directive
for the tests to do so.

I'd prefer this to be not runtime- to push it into some
trick of the type system- but typing.Annotated must be
bound to something and that will lead to muddying up the
scopes of where the directive is placed.  Beyond that,
some of these directives *will be* runtime chained, and
this is intentional to allow a 'break glass' for things
I can't provide a better API for.

Signed-off-by: Brian Harring <ferringb@gmail.com>