chore: rewrite suppress_deprecations to handle generators
Warnings filters are dynamic for the context, and generators resume
into the requesting context- not what they started in.  It's a known thing,
it's a gotcha that's a PITA.

Either I try to explain the problem in the docstring, or just write code
that suppresses it via injecting a generator 'shape' in front of the
actual generator to suppress, and flip the supression on/off around entry
and exit of the generator.

Tests were added, including asserting the python implementation behavior
this suppresses.

Signed-off-by: Brian Harring <ferringb@gmail.com>