chore: fix deprecation markers so they're usable for news, tighten versions.
The large bulk of chunks of this are things that must be resolved in
pkgcore- including optionally transferring it into pkgcore itself so it's
an internal, and can be iterated on faster w/in that.

Effectively lock most of these removes to 0.12.0.  Last released is 0.12.x,
and these changes will go out in 0.11.0 which isn't a semver break,
but large enough changes that it'll be bumped.

Signed-off-by: Brian Harring <ferringb@gmail.com>