Allow deprecations.suppress_deprecations to *not* suppress warnings w/in generators.
Pkgcore's repo.itermatch will be marked deprecated- however both for visibility and
performance, we will *not* want it suppressing deprecations within.

Thus make this configurable.  I'm not sure it'll actually be used there, but
I suspect in downstream consumers of this API- as it's extended into that code-
this will be a necessary edgecase configurable.

Signed-off-by: Brian Harring <ferringb@gmail.com>