pull all contexts.SplitExec usage and contexts.Namespace
Fact is, this may work, but it's deeply intrinsic knowledge of the target
python environment.  I personally can follow the code- and maintain it- but
why *should* snakeoil?  This functionality should only be exposed if we're
leveraging a 3rd party that wants to maintain compatibility for this sort of
intrinsic "sidestep around the python implementation".

This is being pulled due to those concerns, but also since it randomly fails
out in the 0.11.0 release path.  The only change between the last release and
that one for this code is literally running the test from a tarball release
rather than from git- IE, none that could be found, even in replicating in
local envs.  The likely issue here is the pypi env runs in some fashion slightly
different to flex things and expose the issue, but it was reproducible in GH.

So as said: have someone else maintain this since it's not simple, and remove
it until it becomes necessary to have that.

Removing this removes `contexts.Namespace` which runs code w/in a context block
in a namespaced environment.

That's blatantly cool and very ergonomic, but it's built on SplitExec, it's
unused in the ecosystem, so it's being removed in parallel.

This may be revisited; the ergonomic pattern *is* stupidly fluid, but the
maintenance burden has to be less.

Signed-off-by: Brian Harring <ferringb@gmail.com>