reimplement delay import via delayed.import_module
This time around we can do this via just proxying against the
shape of a module, wrapping importlib to do it.  Much, much
simpler since python caught up.

Usage of this to cycle break is a sign of "restructure yo shiat",
but it's also usable for perf reasons.  Much like why python
added lazy imports for 3.15, and this is limited to just module
imports to make that transition simpler.

Signed-off-by: Brian Harring <ferringb@gmail.com>